<hotfix> : Bid 를 가져올때 가격이 Int 의 범위를 넘어가는 버그 픽스

+++ src/main/kotlin/taeyun/malanalter/alertitem/dto/ItemBidInfo.kt
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonCreator
 
 data class ItemBidInfo (
     val tradeStatus: Boolean,
-    val itemPrice: Int,
+    val itemPrice: ULong,
     val comment: String,
     val itemName: String,
     val tradeType: TradeType
